Hello World!
First commit for this project. This is a website for my dad, Eric Kozol,
and his law firm, Kozol Law Offices. It is based off of a theme because
I am lazy and writing this really quickly as a "temporary" site. He owes
me a fancy cider and a gluten free pizza in return for this.